small fix for id of nonetype

+++ sweeperbot/bot.py
@@ -186,22 +186,23 @@ class Bot(commands.AutoShardedBot):
     # little hack to allow only port requests in specified channel
     async def on_message(self, message):
         guild = message.guild
-        settings = self.guild_settings.get(guild.id)
-        request_channel = settings.request_channel
-        if message.channel.id == request_channel:
-            if (
-                message.content[1:].startswith('requestport ')
-                or message.content[1:].startswith('request ')
-                or message.content[1:].startswith('rq ')
-                or message.content[1:].startswith('rqp ')
-                or message.author == self.user
-            ):
-                await self.process_commands(message)
-            else:
-                await message.delete()
-                return
-        else:
-            await self.process_commands(message)
+        if guild:
+          settings = self.guild_settings.get(guild.id)
+          request_channel = settings.request_channel
+          if message.channel.id == request_channel:
+              if (
+                  message.content[1:].startswith('requestport ')
+                  or message.content[1:].startswith('request ')
+                  or message.content[1:].startswith('rq ')
+                  or message.content[1:].startswith('rqp ')
+                  or message.author == self.user
+              ):
+                  await self.process_commands(message)
+              else:
+                  await message.delete()
+                  return
+          else:
+              await self.process_commands(message)
 
     async def on_command_error(self, ctx, error):
         if isinstance(error, commands.NoPrivateMessage):
